Restyle expense list
