+++ src/App.js
@@ -4,9 +4,16 @@ import { getUKFormattedDate, getCurrentMonth } from './utils/helpers';
 import Summary from './views/Summary';
 import AddExpenses from './views/AddExpenses';
 import MonthSelector from './components/MonthSelector';
+import ExpenseHistory from './views/ExpenseHistory';
+import ExpenseAddedAlert from './components/ExpenseAddedAlert';
+import ExpenseDeletedAlert from './components/ExpenseDeletedAlert';
 
 const token = process.env.REACT_APP_MOSTASH_API_KEY;
 const baseURL = process.env.REACT_APP_MOSTASH_BASE_URL;
+const headers = {
+  Stash: token,
+  'Content-Type': 'application/json',
+};
 
 function App() {
   const currentMonth = getCurrentMonth();
@@ -14,6 +21,12 @@ function App() {
   const [selectedMonth, setSelectedMonth] = useState(currentMonth);
   const [expenses, setExpenses] = useState([]);
   const [filteredExpenses, setFilteredExpenses] = useState(expenses);
+
+  // `expenseAdded` and `expenseDeleted` can be undefined, true or false
+  // alerts are hidden when these are undefined
+  const [expenseAdded, setExpenseAdded] = useState();
+  const [expenseDeleted, setExpenseDeleted] = useState();
+  const [lastDeleted, setLastDeleted] = useState({});
 
   useEffect(() => {
     const fetchExpenses = async () => {
@@ -45,6 +58,48 @@ function App() {
     setFilteredExpenses(expensesToRender);
   }, [expenses, selectedMonth]);
 
+  const addExpense = async newExpense => {
+    const url = `${baseURL}/items.json?kind=expense`;
+    const requestOptions = {
+      method: 'POST',
+      headers,
+      body: JSON.stringify(newExpense),
+    };
+
+    // Reset alerts in case user did not dismiss them
+    setExpenseAdded(undefined);
+    setExpenseDeleted(undefined);
+
+    try {
+      const response = await fetch(url, requestOptions);
+      const newExpenseData = await response.json();
+      setExpenses([...expenses, newExpenseData]);
+      setExpenseAdded(true);
+    } catch {
+      setExpenseAdded(false);
+    }
+  };
+
+  const removeExpense = async expense => {
+    const { id } = expense;
+    const url = `${baseURL}/items/${id}.json`;
+    const requestOptions = { method: 'DELETE', headers };
+
+    // Reset alerts in case user did not dismiss them
+    setExpenseAdded(undefined);
+    setExpenseDeleted(undefined);
+
+    try {
+      await fetch(url, requestOptions);
+      setLastDeleted({ ...expense });
+      setExpenses(expenses.filter(expense => expense.id !== id));
+      if (filteredExpenses.length === 1) setSelectedMonth(currentMonth);
+      setExpenseDeleted(true);
+    } catch {
+      setExpenseDeleted(false);
+    }
+  };
+
   return (
     <>
       <header className="navbar bg-dark bg-opacity-10">
@@ -68,14 +123,34 @@ function App() {
           </section>
           <section className="col-lg-6">
             <AddExpenses
-              expenses={expenses}
-              setExpenses={setExpenses}
+              addExpense={addExpense}
+              expenseAdded={expenseAdded}
+              setExpenseAdded={setExpenseAdded}
+              expenseDeleted={setExpenseDeleted}
+              setExpenseDeleted={setExpenseDeleted}
+              lastDeleted={lastDeleted}
+            />
+          </section>
+          <section className="col">
+            <ExpenseHistory
+              removeExpense={removeExpense}
               filteredExpenses={filteredExpenses}
-              setSelectedMonth={setSelectedMonth}
-              currentMonth={currentMonth}
             />
           </section>
         </div>
+        {expenseAdded === undefined ? null : (
+          <ExpenseAddedAlert
+            expenseAdded={expenseAdded}
+            setExpenseAdded={setExpenseAdded}
+          />
+        )}
+        {expenseDeleted === undefined ? null : (
+          <ExpenseDeletedAlert
+            expenseDeleted={expenseDeleted}
+            setExpenseDeleted={setExpenseDeleted}
+            {...lastDeleted}
+          />
+        )}
       </main>
     </>
   );
